fix: move scaffold to repo root, fix fonts + env + security
- Move all nextjs/ files to repo root via git mv (preserves history)
- Replace Geist fonts with Inter + JetBrains Mono in layout.tsx
- Update metadata: title "AI/ML Quest", correct description
- Fix globals.css: use --font-inter/--font-mono, add --border-hover var
- Add .env.example with DATABASE_URL, NextAuth, GitHub OAuth placeholders
- Replace hardcoded passwords in docker-compose.yml with env var refs
- Add Dockerfile from nextjs/ (multi-stage Next.js build, replaces nginx stub)
- Add CHECK constraint comment to drizzle/schema.ts
- Expand .dockerignore to exclude v1 static files from Next.js build
- Merge nextjs/.gitignore into root .gitignore
