+++ .env.example
@@ -1,0 +1,15 @@
+# Database (shared Postgres instance)
+DATABASE_URL=postgresql://user:password@localhost:5432/dbname
+
+# Postgres credentials (for local docker-compose only)
+POSTGRES_USER=crawler_admin
+POSTGRES_PASSWORD=change-me
+POSTGRES_DB=n8n
+
+# NextAuth.js
+NEXTAUTH_SECRET=generate-with-openssl-rand-base64-32
+NEXTAUTH_URL=http://localhost:3003
+
+# GitHub OAuth (create at https://github.com/settings/developers)
+GITHUB_ID=
+GITHUB_SECRET=
