+++ Dockerfile
@@ -1,3 +1,38 @@
-FROM nginx:alpine
-COPY . /usr/share/nginx/html/
-EXPOSE 80
+# Multi-stage build for Next.js 15 standalone output
+# Targets port 3003, deployed alongside other services on shared Postgres
+
+# ── Stage 1: Builder ────────────────────────────────────────────────────────
+FROM node:20-alpine AS builder
+
+WORKDIR /app
+
+COPY package*.json ./
+RUN npm ci
+
+COPY . .
+
+# Build requires DATABASE_URL at build time (for type generation only, not secrets)
+# In CI/CD, pass as build arg if needed. For local: use .env.local
+ARG DATABASE_URL=postgresql://placeholder:placeholder@localhost:5432/placeholder
+ENV DATABASE_URL=${DATABASE_URL}
+
+RUN npm run build
+
+# ── Stage 2: Runner ─────────────────────────────────────────────────────────
+FROM node:20-alpine AS runner
+
+WORKDIR /app
+
+ENV NODE_ENV=production
+
+# Copy standalone output
+COPY --from=builder /app/.next/standalone ./
+COPY --from=builder /app/.next/static ./.next/static
+COPY --from=builder /app/public ./public
+
+EXPOSE 3003
+
+ENV PORT=3003
+ENV HOSTNAME="0.0.0.0"
+
+CMD ["node", "server.js"]
