+++ docker-compose.yml
@@ -7,10 +7,11 @@ services:
     build: .
     ports:
       - "3003:3003"
+    env_file: .env
     environment:
-      - DATABASE_URL=postgresql://crawler_admin:password@postgres:5432/n8n
-      - NEXTAUTH_SECRET=dev-secret-change-in-production
-      - NEXTAUTH_URL=http://localhost:3003
+      - DATABASE_URL=${DATABASE_URL}
+      - NEXTAUTH_SECRET=${NEXTAUTH_SECRET:-dev-secret}
+      - NEXTAUTH_URL=${NEXTAUTH_URL:-http://localhost:3003}
     depends_on:
       postgres:
         condition: service_healthy
@@ -19,16 +20,20 @@ services:
   postgres:
     image: postgres:16-alpine
     environment:
-      POSTGRES_USER: crawler_admin
-      POSTGRES_PASSWORD: password
-      POSTGRES_DB: n8n
+      POSTGRES_USER: ${POSTGRES_USER:-crawler_admin}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+      POSTGRES_DB: ${POSTGRES_DB:-n8n}
     ports:
       - "5432:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
       - ../postgres/migrations:/docker-entrypoint-initdb.d
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U crawler_admin -d n8n"]
+      test:
+        [
+          "CMD-SHELL",
+          "pg_isready -U ${POSTGRES_USER:-crawler_admin} -d ${POSTGRES_DB:-n8n}",
+        ]
       interval: 5s
       timeout: 5s
       retries: 5
